requirements.txt
# Core ML
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
accelerate>=0.33.0
transformers>=4.44.0
diffusers>=0.30.0
peft>=0.12.0
safetensors>=0.4.0

# Quantization
bitsandbytes>=0.49.0

# Video processing
opencv-python-headless>=4.8.0
ffmpeg-python>=0.2.0
scenedetect[opencv]>=0.6.4

# Captioning
qwen-vl-utils>=0.0.8

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
pillow>=10.0.0
einops>=0.7.0
numpy>=1.24.0
psutil>=5.9.0
